Move geneti parameters to GPGraphOptimizerParameters
